Use LC::Stat constants in ccm-purge
The previous attempt to convert ccm-purge to use atime instead of mtime
of the old profiles update the wrong part of the code. Having tested
updating the correct part of the code, it seems that ccm-purge can
modify the atime of the profile directory which is perhaps why it was
using mtime. The use of mod_time as the variable name also suggeststhat
perhaps the documentation was wrong not the code.

Updated the documentation to reflect the use of mtime and also
introduced the use of LC::Stat's constants to make it more readable.