CacheManager::DB: set CLOEXEC on tied DB files (for DB_File and CDB_File, not supported for GDBM)
File descriptors are passed to child processes on POSIX platforms by design,
in the case of CCM, this can lead to these being passed all the way down the call tree
to processes (e.g. daemons) that are started by components.
Setting CLOEXEC disables this behaviour and fixes the potential security risk that this poses.

This problem is tackled in modern OSes in e.g. glibc or even perl itself when using regular open.
However, this is still an issue for the CCM DB files, since we pass our own flags, bypassing
the default behaviour.