Strip out all references to derivation
Remove all references to this functionality as it is not supported.

This appears to have been a method for tracking where profile values were changed but was removed after the very first Quattor workshop for performance reasons and was never implemented in the current Java implementation version of panc.

This is backwards incompatible as it changes the on-disk DB format, but the field has always been null.
Add `ccm fetch -f` to the post-install script to ensure the profile is updated immediately.