ProfileCache: Include type in checksum calculation
The goal here is to eventually be able to recompute a checksum for the whole profile to verify integrity,
this requires the checksums to include type so that alteration or corruption of type can be detected.

N.B. \x1e is the record separator character, used here to delimit the value and type.